+++ CHANGELOG.rst
@@ -4,9 +4,11 @@ Change Log
 `X.Y.Z`_ (TBD-DD-DD)
 ---------------------
 
-* Add support for PostgreSQL 13 and 14 RDS parameter groups
-* Drop support for RDS PostgreSQL 9.x
 * Add support for T4g instance types
+* Add support for PostgreSQL 13 and 14 RDS parameter groups
+* Add support for `EKS EncryptionConfig <https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-eks-cluster.html#cfn-eks-cluster-encryptionconfig>`_. Set with ``EnableEksEncryptionConfig=true``.
+* Add ``EksClusterName`` parameter to control name of EKS cluster. If upgrading, set this to STACK_NAME-cluster to match existing name.
+* Drop support for RDS PostgreSQL 9.x
 
 
 `2.1.2`_ (2022-03-10)